task 4 part 2 is done

+++ five_all_tasks/task4/Bill.java
@@ -1,11 +1,13 @@
 package five_all_tasks.task4;
 
+import java.math.BigDecimal;
+
 public class Bill {
-    private double amount;
+    private BigDecimal amount;
     private TaxType taxType;
     private TaxService taxService;
 
-    public Bill(double amount, TaxType taxType, TaxService taxService) {
+    public Bill(BigDecimal amount, TaxType taxType, TaxService taxService) {
         this.amount = amount;
         this.taxType = taxType;
         this.taxService = taxService;
@@ -14,7 +16,7 @@ public class Bill {
     public void payTaxes() {
         // TODO вместо 0.0 посчитать размер налога исходя из TaxType
 
-        double taxAmount = taxType.calculateTaxFor(amount);
+        BigDecimal taxAmount = taxType.calculateTaxFor(amount);
         taxService.payOut(taxAmount);
     }
 }
