+++ five_all_tasks/task4/TaxService.java
@@ -1,7 +1,9 @@
 package five_all_tasks.task4;
 
+import java.math.BigDecimal;
+
 public class TaxService {
-    public void payOut(double taxAmount) {
+    public void payOut(BigDecimal taxAmount) {
         System.out.format("Уплачен налог в размере %.2f%n", taxAmount);
     }
 }
